docwriter/pyproject.toml
[project]
name = "docwriter"
version = "0.1.0"
description = ""
authors = [
    {name = "ao32v",email = "ao32v@cummins.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "mkdocs-material (>=9.6.22,<10.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ruamel-yaml (>=0.18.16,<0.19.0)",
    "pathvalidate (>=3.3.1,<4.0.0)",
    "pyside6 (>=6.10.0,<7.0.0)"
]

[tool.poetry]
packages = [{include = "docwriter", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
